Organized Frontend
Added Folders for each Components and split .ts into .html .scss and .ts for more clarity
