+++ frontend/src/app/admin-dashboard/admin-dashboard.component.html
@@ -132,4 +132,4 @@
 
 <ng-template #accessDenied>
   <p style="color: red;">Zugriff verweigert. Sie müssen ein Dozent sein, um diese Seite anzuzeigen.</p>
-</ng-template>
+</ng-template>
