+++ frontend/src/app/admin-dashboard/admin-dashboard.component.scss
@@ -1,0 +1,5 @@
+.actions-panel {
+  margin-bottom: 20px;
+  display: flex;
+  gap: 10px;
+}
