+++ frontend/src/app/app.component.ts
@@ -5,7 +5,7 @@ import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatMenuModule } from '@angular/material/menu';
-import { BackendStatusComponent } from './backend-status.component';
+import { BackendStatusComponent } from './backend-status/backend-status.component';
 import { AuthService } from './auth.service';
 
 @Component({
