+++ frontend/src/app/backend-status/backend-status.component.html
@@ -2,4 +2,4 @@
   <span *ngIf="status === 'up'" class="status-text">SQL Platform API bereit</span>
   <span *ngIf="status === 'down'" class="status-text">SQL Platform API nicht erreichbar</span>
   <span *ngIf="status === 'unknown'" class="status-text">Prüfe API Status...</span>
-</div>
+</div>
