+++ frontend/src/app/backend-status/backend-status.component.scss
@@ -8,4 +8,4 @@
 }
 .unknown .status-text {
   color: #888;
-}
+}
