+++ frontend/src/app/exercises/exercises.component.scss
@@ -1,0 +1,27 @@
+.exercises-container {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 20px;
+  margin-top: 20px;
+}
+
+.exercise-card {
+  width: 300px;
+  margin-bottom: 20px;
+}
+
+.no-exercises-card {
+  width: 100%;
+  text-align: center;
+  padding: 30px;
+  background-color: #f5f5f5;
+}
+
+h1 {
+  color: #3f51b5;
+  margin-bottom: 20px;
+}
+
+.error-message {
+  color: red;
+}
