+++ frontend/src/app/exercises/exercises.component.ts
@@ -1,0 +1,16 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { AuthService } from '../auth.service';
+import { MatCardModule } from '@angular/material/card';
+import { MatButtonModule } from '@angular/material/button';
+
+@Component({
+  selector: 'app-exercises',
+  standalone: true,
+  imports: [CommonModule, MatCardModule, MatButtonModule],
+  templateUrl: './exercises.component.html',
+  styleUrls: ['./exercises.component.scss']
+})
+export class ExercisesComponent {
+  constructor(public authService: AuthService) {}
+}
