--- frontend/src/app/info-box/info-box.component.html
@@ -1,3 +1,0 @@
-<p>info-box works!</p>
-<button (click)="hidden=!hidden">Toggle</button>
-<p [hidden]="hidden">{{text}}</p>
