--- frontend/src/app/info-box/info-box.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { InfoBoxComponent } from './info-box.component';
-
-describe('InfoBoxComponent', () => {
-  let component: InfoBoxComponent;
-  let fixture: ComponentFixture<InfoBoxComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [InfoBoxComponent]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(InfoBoxComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
