+++ frontend/src/app/profile/profile.component.html
@@ -57,4 +57,4 @@
     <mat-icon>error</mat-icon>
     <span>{{error}}</span>
   </div>
-</div>
+</div>
