+++ frontend/src/app/register/register.component.scss
@@ -1,0 +1,34 @@
+:host {
+  display: block;
+}
+
+mat-card {
+  max-width: 400px;
+  margin: 2em auto;
+}
+
+h2 {
+  text-align: center;
+}
+
+mat-form-field {
+  width: 100%;
+  margin-bottom: 1em;
+}
+
+.actions {
+  text-align: center;
+  margin-top: 1.5em;
+}
+
+.error-message {
+  color: #c62828;
+  text-align: center;
+  margin-top: 1em;
+}
+
+.success-message {
+  color: #388e3c;
+  text-align: center;
+  margin-top: 1em;
+}
