+++ frontend/src/app/teacher-dashboard/teacher-dashboard.component.html
@@ -1,0 +1,45 @@
+<div *ngIf="authService.isTeacher() || authService.isTutor(); else accessDenied">
+  <h1>Dozenten-Dashboard</h1>
+  
+  <div class="dashboard-container">
+    <mat-card class="dashboard-card">
+      <mat-card-header>
+        <mat-card-title>Übungsverwaltung</mat-card-title>
+      </mat-card-header>
+      <mat-card-content>
+        <p>Erstellen und verwalten Sie SQL-Übungen für Ihre Studenten.</p>
+      </mat-card-content>
+      <mat-card-actions>
+        <button mat-button color="primary" routerLink="/teacher/exercises">ÜBUNGEN VERWALTEN</button>
+      </mat-card-actions>
+    </mat-card>
+    
+    <mat-card class="dashboard-card">
+      <mat-card-header>
+        <mat-card-title>Studentenfortschritt</mat-card-title>
+      </mat-card-header>
+      <mat-card-content>
+        <p>Verfolgen Sie den Fortschritt der Studenten bei den Übungen.</p>
+      </mat-card-content>
+      <mat-card-actions>
+        <button mat-button color="primary">FORTSCHRITT ANZEIGEN</button>
+      </mat-card-actions>
+    </mat-card>
+    
+    <mat-card class="dashboard-card">
+      <mat-card-header>
+        <mat-card-title>Datenbankschemata</mat-card-title>
+      </mat-card-header>
+      <mat-card-content>
+        <p>Verwalten Sie Datenbankschemata für Übungen.</p>
+      </mat-card-content>
+      <mat-card-actions>
+        <button mat-button color="primary">SCHEMATA VERWALTEN</button>
+      </mat-card-actions>
+    </mat-card>
+  </div>
+</div>
+
+<ng-template #accessDenied>
+  <p style="color: red;">Zugriff verweigert. Sie müssen ein Dozent oder Tutor sein, um diese Seite anzuzeigen.</p>
+</ng-template>
