+++ frontend/src/app/teacher-dashboard/teacher-dashboard.component.scss
@@ -1,0 +1,16 @@
+.dashboard-container {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 20px;
+  margin-top: 20px;
+}
+
+.dashboard-card {
+  width: 300px;
+  margin-bottom: 20px;
+}
+
+h1 {
+  color: #3f51b5;
+  margin-bottom: 20px;
+}
