+++ frontend/src/app/teacher-dashboard/teacher-dashboard.component.ts
@@ -1,0 +1,17 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { AuthService } from '../auth.service';
+import { MatCardModule } from '@angular/material/card';
+import { MatButtonModule } from '@angular/material/button';
+import { RouterModule } from '@angular/router';
+
+@Component({
+  selector: 'app-teacher-dashboard',
+  standalone: true,
+  imports: [CommonModule, MatCardModule, MatButtonModule, RouterModule],
+  templateUrl: './teacher-dashboard.component.html',
+  styleUrls: ['./teacher-dashboard.component.scss']
+})
+export class TeacherDashboardComponent {
+  constructor(public authService: AuthService) {}
+}
