+++ frontend/src/app/teacher-exercises/teacher-exercises.component.scss
@@ -1,0 +1,35 @@
+.exercises-container {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 20px;
+  margin-top: 20px;
+}
+
+.exercise-card {
+  width: 400px;
+  margin-bottom: 20px;
+}
+
+.no-exercises-card {
+  width: 100%;
+  text-align: center;
+  padding: 30px;
+  background-color: #f5f5f5;
+}
+
+h1 {
+  color: #3f51b5;
+  margin-bottom: 20px;
+}
+
+.actions {
+  margin-bottom: 20px;
+}
+
+.stats {
+  display: flex;
+  justify-content: space-between;
+  margin-top: 15px;
+  padding-top: 10px;
+  border-top: 1px solid #eee;
+}
